move vcs hook
